split merging into 2 different python files
